refactor: moved cmd to separate files
