+++ modules/handlers/__init__.py
@@ -1,5 +1,8 @@
 """Modules that handle the events the bot recognizes and reacts to"""
 from modules.data import config_map
+from .start import start_cmd
+from .help import help_cmd
+from .rules import rules_cmd
 
 CHAT_PRIVATE_ERROR = f"Non puoi usare quest comando ora\nChatta con {config_map['bot_tag']} in privato"
 INVALID_MESSAGE_TYPE_ERROR = "Questo tipo di messaggio non è supportato\n"\
